+++ Tests/Get-StuccoTemplate.tests.ps1
@@ -1,0 +1,6 @@
+describe 'Get-StuccoTemplate' {
+    it 'Returns the Stucco template' {
+        $t = Get-StuccoTemplate
+        $t.psobject.TypeNames.Contains('System.Management.Automation.PSCustomObject') | Should -Be $true
+    }
+}